# account_service/swagger/account_service_api.yaml
swagger: "2.0"
info:
  title: Dukedoms of Daleria Account Service API
  version: "0.1.2"

paths:
  /newAccount:
    post:
      summary: create new player account
      x-swagger-router-controller: controllers.account_creation
      operationId: create_new_account
      tags:
        - newAccount
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: newAccountRequest
          description: request to create new account
          schema:
            $ref: '#/definitions/NewAccountRequest'
      responses:
        200:
          description: Player Account Successfully Created
          schema:
            $ref: '#/definitions/NewAccountRequestSuccessful'

  /getAccountInfo/{accountId}:
    get:
      summary: get all information for given playerId
      x-swagger-router-controller: controllers.account_info
      operationId: get_player_info
      tags:
        - accountInfo
      parameters:
        - in: path
          name: accountId
          type: integer
          required: true
          description: numeric id of player whose info to fetch
      responses:
        200:
          description: Player account found successfully
          schema:
            $ref: '#/definitions/accountInfo'
        404:
          description: Player id not found
  /getAccountIds:
    get:
      summary: get ids for submitted accounts
      x-swagger-router-controller: controllers.account_info
      operationId: get_account_ids
      tags:
        - accountInfo
      parameters:
        - in: query
          name: requestedAccounts
          type: array
          items:
            type: string
          required: true
      responses:
        200:
          description: all accounts found
          schema:
            $ref: '#/definitions/AccountIdMappings'
        404:
          description: at least one provided account not found

  /verifyAccounts:
    get:
      summary: determine if a list of accountIds exist in the system
      x-swagger-router-controller: controllers.account_info
      operationId: verify_accounts
      tags:
        - accountInfo
      parameters:
        - in: query
          name: playerList
          type: array
          items:
            type: string
          required: true
      responses:
        200:
          description: successful
          schema:
            $ref: '#/definitions/UnverifiedPlayers'

  /invitePlayers:
    post:
      summary: invite set of players to game
      x-swagger-router-controller: controllers.gameOperations
      operationId: invite_accounts
      tags:
        - gameOperations
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: invitationBatch
          schema:
            $ref: '#/definitions/InvitationBatch'
      responses:
        202:
          description: invitations successfully submitted
        404:
          description: at least one player not found


definitions:
  NewAccountRequest:
    type: object
    properties:
      email:
        type: string
    required:
      - email
  NewAccountRequestSuccessful:
    type: object
    properties:
      email:
        type: string
      accountId:
        type: integer
  accountInfo:
    type: object
    properties:
      email:
        type: string
      accountId:
        type: integer
      activePlayerIds:
        type: object
      pendingPlayerIds:
        type: object
      gameInvitations:
        type: array
        items:
          type: integer
  UnverifiedPlayers:
    type: object
    properties:
      UnverifiedPlayers:
        type: array
        items:
          type: string
  AccountIdMappings:
    type: object
    properties:
      accountIdMappings:
        type: array
        items:
          type: object
  InvitationBatch:
    type: object
    properties:
      gameId:
        type: integer
      playerList:
        type: array
        items:
          type: string